feat: move projection head tests to tailor
